requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pinecone-client==2.2.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
segment-anything==1.0
six==1.16.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0
